third_party/WebKit/LayoutTests/inspector/runtime/runtime-localStorage-getProperties.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script>

localStorage.testProperty = "testPropertyValue";

async function test()
{
    var result = await InspectorTest.RuntimeAgent.evaluate("localStorage");
    var localStorageHandle = InspectorTest.runtimeModel.createRemoteObject(result);
    localStorageHandle.getOwnProperties(false, step2);

    function step2(properties)
    {
        for (var property of properties) {
            if (property.name !== "testProperty")
                continue;
            property.value = { type: property.value.type, description: property.value.description };
            InspectorTest.dump(property);
        }
        InspectorTest.completeTest();
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests RemoteObject.getProperties on localStorage object. <a href="https://bugs.webkit.org/show_bug.cgi?id=66215">66215</a>
</p>

</body>
</html>
