third_party/WebKit/LayoutTests/inspector/runtime/runtime-localStorage-getProperties-expected.txt

Tests RemoteObject.getProperties on localStorage object. 66215

{
    enumerable : true
    name : "testProperty"
    value : {
        description : "testPropertyValue"
        type : "string"
    }
    writable : true
}
